agentic_ai_fund_manager/financial_analyst/requirements.txt
# requirements.txt
# Financial Analyst Dependencies

# AI Agent Framework
strands-agents
strands-agents-tools

# AWS Bedrock AgentCore
bedrock-agentcore
